Move js integration utils to a crate!
