+++ Cargo.toml
@@ -3,6 +3,7 @@
 members = [
   "hdht",
   "rpc",
-  "swarm"
+  "swarm",
+  "test_utils"
 ]
 resolver = "2"
