+++ hdht/Cargo.toml
@@ -46,6 +46,8 @@ serde_json = "1.0.135"
 tempfile = "3.10.1"
 tracing-subscriber = { version = "0.3.19", features = ["env-filter"] }
 
+[dev-dependencies.test_utils]
+path = "../test_utils"
 
 [dev-dependencies.rusty_nodejs_repl]
 features = [ "serde" , "integration_utils"]
