+++ swarm/Cargo.toml
@@ -13,9 +13,7 @@ tracing = "0.1.41"
 rand = "0.7.3"
 
 [dev-dependencies]
+test_utils = { path = "../test_utils" }
 tokio = { version = "1.41.0", features = ["rt-multi-thread", "macros", "time"] }
 tracing-subscriber = { version = "0.3.19", features = ["env-filter"] }
-rusty_nodejs_repl = { version = "0.2.2", features = ["serde", "integration_utils"] }
-serde = "1.0.217"
-serde_json = "1.0.135"
 hypercore_handshake = "0.3.0"
