+++ swarm/tests/integration.rs
@@ -1,12 +1,10 @@
 //! Integration tests for hyperswarm
 
-mod common;
-
-use common::{Result, Testnet};
 use dht_rpc::IdBytes;
 use futures::{SinkExt, StreamExt, join};
 use hypercore_handshake::CipherEvent;
 use hyperswarm::{DhtConfig, JoinOpts, Swarm, SwarmConfig};
+use test_utils::{Result, Testnet, log};
 
 /// Server announces and client discovers via lookup
 #[tokio::test]
