+++ test_utils/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "test_utils"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+async-process = "2.2.1"
+rusty_nodejs_repl = { version = "0.2.2", features = ["serde", "integration_utils"] }
+serde = "1.0.217"
+serde_json = "1.0.135"
+tempfile = "3.10.1"
+thiserror = "1.0.68"
+tracing = "0.1.41"
+tracing-subscriber = { version = "0.3.19", features = ["env-filter"] }
